This commit has done the following:  + Verified that the routing table is populated in its entirety from the testbench.  + Verified that the routing address is assigned to the router (using 0101 as the test address) from the testbench  + Made a solid attempt at injecting a control packet from the North port (source: 0001). Here is how the router reacted:       - The router sensed the CTRLFlg toggled from the toplevel of BlazeRouter_TB       - Read in the control packet off the north port       - Set the CTR flag back to the faux neighbor, subsequently causing the testbench to see the change and update the CTRLflg       - Registered the control packet successfully in the reservation, scheduler, and ID table.       - Started the scheduler and ran for the specified time (1205 cycles)       - Expired on time and departed some empty data within the North Cell.
Note: No data packet was submitted to the router in this test. This will be addressed in the next commit.
Corrections: The testbench wasn't updated properly to reflect the intended direction (DIR) that the control packet was reserving. It was supposed to be West but was left as North, and hence upon departure, the arbiter accessed the VCC and set the switch to North. I corrected this, but I will test it in the next commit.


Signed-off-by: koru <krikorh@unlv.nevada.edu>